backend-python/requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
redis>=5.0.0
sentence-transformers>=2.2.0  # For later tasks
faiss-cpu>=1.7.4             # For later tasks
pymongo>=4.4.0
motor>=3.2.0